Improve robustness, safety, and tests (#86)
Multiple fixes and improvements across the codebase: make several OllamaClient helpers package-visible for testing; pass AI request timeout from settings when pinging; harden DbSeedSettingsState by copying maps, returning unmodifiable views and enforcing defaults; make PendingUpdate and Row defensively copy/immut immutable collections; guard ProgressTracker against non-positive advances; preserve decimal bounds in SchemaIntrospector by switching min/max to Double; expand SQL reserved keyword set; switch TopologicalSorter to an iterative Tarjan implementation to avoid stack overflows and added a deep-chain test; make array formatting configurable and clean control characters in AbstractDialect, plus batch size guard; add DialectFactory caching; add logging and safer handling in ForeignKeyResolver (depth normalization, non-nullable-cycle warnings, changed behavior to defer/return null when appropriate) and add related tests; RowGenerator logs when it cannot generate full row counts and parses trimmed values; ValueGenerator and Column updated to use Double for min/max and convert to int as needed; RepetitionRule validates count >= 1; simplify RepetitionRulesPanel UI by removing prefix handling; DriverLoader centralizes driver dir handling, ensures directory creation, adds deregisterAll and classloader cleanup; NotificationHelper uses central notification constant; and many unit tests updated/added to cover these behaviors. These changes improve safety, immutability, configurability, and test coverage.

+++ src/main/java/com/luisppb16/dbseed/db/SchemaIntrospector.java
@@ -201,13 +201,13 @@ public class SchemaIntrospector {
     final Set<String> pkSet = new LinkedHashSet<>(pkCols);
 
     for (final ColumnRawData raw : rawColumns) {
-      Integer minValue = null;
-      Integer maxValue = null;
+      Double minValue = null;
+      Double maxValue = null;
 
       final double[] bounds = inferBoundsFromChecks(checkConstraints, raw.name());
       if (bounds.length == 2) {
-        minValue = (int) bounds[0];
-        maxValue = (int) bounds[1];
+        minValue = bounds[0];
+        maxValue = bounds[1];
       }
 
       final Set<String> allowedValues = inferAllowedValuesFromChecks(checkConstraints, raw.name());
